text update for my stories
