+++ assets/src/dashboard/constants/index.js
@@ -91,19 +91,6 @@ export const secondaryPaths = [
   },
 ];
 
-export const STORY_STATUS = {
-  ALL: 'publish,draft',
-  PUBLISHED: 'publish',
-  DRAFT: 'draft',
-};
-
-export const STORY_STATUSES = [
-  { label: __('All Stories', 'web-stories'), value: STORY_STATUS.ALL },
-  { label: __('Drafts', 'web-stories'), value: STORY_STATUS.DRAFT },
-  { label: __('Active Stories', 'web-stories'), value: STORY_STATUS.PUBLISHED },
-  // { label: __('My Templates', 'web-stories'), value: 'template ' },
-];
-
 export const VIEW_STYLE = {
   GRID: 'GRID',
   LIST: 'LIST',
@@ -115,91 +102,9 @@ export const ICON_METRICS = {
   LEFT_RIGHT_ARROW: { width: 16, height: 16 },
 };
 
-export const STORY_CONTEXT_MENU_ACTIONS = {
-  OPEN_IN_EDITOR: 'open-in-editor-action',
-  PREVIEW: 'preview-action',
-  RENAME: 'rename-action',
-  DUPLICATE: 'duplicate-action',
-  CREATE_TEMPLATE: 'create-template-action',
-  DELETE: 'delete-story-action',
-};
-
-export const STORY_CONTEXT_MENU_ITEMS = [
-  {
-    label: __('Open in editor', 'web-stories'),
-    value: STORY_CONTEXT_MENU_ACTIONS.OPEN_IN_EDITOR,
-  },
-  {
-    label: __('Preview', 'web-stories'),
-    value: STORY_CONTEXT_MENU_ACTIONS.PREVIEW,
-  },
-  { label: null, value: false, separator: true },
-  {
-    label: __('Rename', 'web-stories'),
-    value: STORY_CONTEXT_MENU_ACTIONS.RENAME,
-  },
-  {
-    label: __('Duplicate', 'web-stories'),
-    value: STORY_CONTEXT_MENU_ACTIONS.DUPLICATE,
-  },
-  {
-    label: __('Create Template', 'web-stories'),
-    value: STORY_CONTEXT_MENU_ACTIONS.CREATE_TEMPLATE,
-  },
-  { label: null, value: false, separator: true },
-  {
-    label: __('Delete Story', 'web-stories'),
-    value: STORY_CONTEXT_MENU_ACTIONS.DELETE,
-  },
-];
-
 export const ITEMS_PER_PAGE = 10;
-
-export const STORY_SORT_OPTIONS = {
-  LAST_MODIFIED: 'modified',
-  LAST_OPENED: false,
-  DATE_CREATED: 'date',
-  CREATED_BY: 'story_author',
-  NAME: 'title',
-};
-
-export const SORT_DIRECTION = {
-  ASC: 'asc',
-  DESC: 'desc',
-};
-
-export const ORDER_BY_SORT = {
-  [STORY_SORT_OPTIONS.LAST_MODIFIED]: SORT_DIRECTION.DESC,
-  [STORY_SORT_OPTIONS.LAST_OPENED]: SORT_DIRECTION.DESC,
-  [STORY_SORT_OPTIONS.DATE_CREATED]: SORT_DIRECTION.DESC,
-  [STORY_SORT_OPTIONS.CREATED_BY]: SORT_DIRECTION.ASC,
-  [STORY_SORT_OPTIONS.NAME]: SORT_DIRECTION.ASC,
-};
-
-export const STORY_SORT_MENU_ITEMS = [
-  {
-    label: __('Last modified', 'web-stories'), // default
-    value: STORY_SORT_OPTIONS.LAST_MODIFIED,
-  },
-  {
-    label: __('Last opened', 'web-stories'),
-    value: STORY_SORT_OPTIONS.LAST_OPENED,
-  },
-  {
-    label: __('Date created', 'web-stories'),
-    value: STORY_SORT_OPTIONS.DATE_CREATED,
-  },
-
-  {
-    label: __('Author', 'web-stories'),
-    value: STORY_SORT_OPTIONS.CREATED_BY,
-  },
-  {
-    label: __('Title', 'web-stories'),
-    value: STORY_SORT_OPTIONS.NAME,
-  },
-];
 
 export * from './animation';
 export * from './direction';
+export * from './stories';
 export * from './templates';
